Extract and rename the easy to use interface 'CoSy' to 'Maestro'
