+++ benchmarks/test_benchmark_maximal_elements.py
